Fixes from CR.
