+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/app_insights/app_insights_datasource.test.ts
@@ -409,55 +409,6 @@ describe('AppInsightsDatasource', () => {
           return Promise.resolve(response);
         });
       });
-
-      it('should return a list of metric names', () => {
-        return ctx.ds.metricFindQueryInternal('appInsightsMetricNames()').then((results: any) => {
-          expect(results.length).toBe(2);
-          expect(results[0].text).toBe('exceptions/server');
-          expect(results[0].value).toBe('exceptions/server');
-          expect(results[1].text).toBe('requests/count');
-          expect(results[1].value).toBe('requests/count');
-        });
-      });
-    });
-
-    describe('with metadata group by query', () => {
-      const response = {
-        metrics: {
-          'exceptions/server': {
-            supportedAggregations: ['sum'],
-            supportedGroupBy: {
-              all: ['client/os', 'client/city', 'client/browser'],
-            },
-            defaultAggregation: 'sum',
-          },
-          'requests/count': {
-            supportedAggregations: ['avg', 'sum', 'total'],
-            supportedGroupBy: {
-              all: ['client/os', 'client/city', 'client/browser'],
-            },
-            defaultAggregation: 'avg',
-          },
-        },
-      };
-
-      beforeEach(() => {
-        ctx.ds.getResource = jest.fn().mockImplementation((path) => {
-          expect(path).toContain('/metrics/metadata');
-          return Promise.resolve(response);
-        });
-      });
-
-      it('should return a list of group bys', () => {
-        return ctx.ds.metricFindQueryInternal('appInsightsGroupBys(requests/count)').then((results: any) => {
-          expect(results[0].text).toContain('client/os');
-          expect(results[0].value).toContain('client/os');
-          expect(results[1].text).toContain('client/city');
-          expect(results[1].value).toContain('client/city');
-          expect(results[2].text).toContain('client/browser');
-          expect(results[2].value).toContain('client/browser');
-        });
-      });
     });
   });
