+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/datasource.ts
@@ -12,9 +12,8 @@ import {
   DataSourceInstanceSettings,
   LoadingState,
   ScopedVars,
-  toDataFrame,
 } from '@grafana/data';
-import { forkJoin, Observable, of, from } from 'rxjs';
+import { forkJoin, Observable, of } from 'rxjs';
 import { getTemplateSrv, TemplateSrv } from '@grafana/runtime';
 import InsightsAnalyticsDatasource from './insights_analytics/insights_analytics_datasource';
 import { datasourceMigrations } from './utils/migrateQuery';
@@ -22,7 +21,7 @@ import { map } from 'rxjs/operators';
 import AzureResourceGraphDatasource from './azure_resource_graph/azure_resource_graph_datasource';
 import { getAzureCloud } from './credentials';
 import migrateAnnotation from './utils/migrateAnnotation';
-import { VariableSupport } from './components/VariableEditor/CustomVariableSupport';
+import { VariableSupport } from './variables';
 export default class Datasource extends DataSourceApi<AzureMonitorQuery, AzureDataSourceJsonData> {
   annotations = {
     prepareAnnotation: migrateAnnotation,
@@ -76,16 +75,6 @@ export default class Datasource extends DataSourceApi<AzureMonitorQuery, AzureDa
     this.variables = new VariableSupport(this);
   }
 
-  queryForGrafanaTemplateVariableFn(query: string) {
-    const promisedResults = async () => {
-      const templateVariablesResults = await this.azureMonitorDatasource.metricFindQueryInternal(query);
-      return {
-        data: templateVariablesResults ? [toDataFrame(templateVariablesResults)] : [],
-      };
-    };
-    return from(promisedResults());
-  }
-
   query(options: DataQueryRequest<AzureMonitorQuery>): Observable<DataQueryResponse> {
     const byType = new Map<AzureQueryType, DataQueryRequest<AzureMonitorQuery>>();
 
@@ -111,11 +100,6 @@ export default class Datasource extends DataSourceApi<AzureMonitorQuery, AzureDa
     }
 
     const observables: Array<Observable<DataQueryResponse>> = Array.from(byType.entries()).map(([queryType, req]) => {
-      // TODO: in the future consider if this should also be it's own pseudoDatasource?
-      if (queryType === AzureQueryType.GrafanaTemplateVariableFn) {
-        return this.queryForGrafanaTemplateVariableFn(req.targets[0].grafanaTemplateVariableFn?.query || '');
-      }
-
       const ds = this.pseudoDatasource[queryType];
       if (!ds) {
         throw new Error('Data source not created for query type ' + queryType);
@@ -149,29 +133,6 @@ export default class Datasource extends DataSourceApi<AzureMonitorQuery, AzureDa
 
   async annotationQuery(options: any) {
     return this.azureLogAnalyticsDatasource.annotationQuery(options);
-  }
-
-  async metricFindQuery(query: string, optionalOptions?: unknown) {
-    if (!query) {
-      return Promise.resolve([]);
-    }
-
-    const aiResult = this.appInsightsDatasource?.metricFindQueryInternal(query);
-    if (aiResult) {
-      return aiResult;
-    }
-
-    const amResult = this.azureMonitorDatasource.metricFindQueryInternal(query);
-    if (amResult) {
-      return amResult;
-    }
-
-    const alaResult = this.azureLogAnalyticsDatasource.metricFindQueryInternal(query, optionalOptions);
-    if (alaResult) {
-      return alaResult;
-    }
-
-    return Promise.resolve([]);
   }
 
   async testDatasource(): Promise<DatasourceValidationResult> {
